requirements.txt
aiohttp==3.7.4.post0
astroid==2.6.5
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
discord.py==1.7.3
idna==3.2
isort==5.9.2
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
multidict==5.1.0
pep8==1.7.1
pkg-resources==0.0.0
pycodestyle==2.7.0
pylint==2.9.5
python-dotenv==0.19.0
requests==2.26.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
yarl==1.6.3
